# Move division by 0 check out of the bloom loops