@aaronbuchwald 2024-01-17
Great meme